+++ oservi/__init__.py
@@ -66,3 +66,4 @@ __all__ = [
 ]
 
 from oservi.engines.channel_watcher import ChannelWatcherEngine
+from oservi.engines.arxiv_watcher import ArxivWatcherEngine